obs_shards/backend/requirements.txt
# Core
Django>=5.0
djangorestframework>=3.15
django-filter>=24.2
dj-database-url
python-dotenv

# Environment
python-dotenv>=1.0

# Database (future-ready)
psycopg[binary]>=3.1

# Utilities
pytz
python-dateutil

# Dev / Quality
black
isort
ruff
mypy
django-stubs

# API docs (future)
drf-spectacular

# Background tasks (future)
celery
redis

# Data / Analysis (future)
pandas
numpy
